refactor file storage controller into separate controllers
- Removed FileStorageController and replaced with StorageFileController,
StorageFolderController, StorageMainController, and
StorageSearchController.
- Moved file and folder related operations into respective controllers.
- Added new utility classes for disposition header building and name
validation.
- Updated HTML templates to use new controller endpoints.
